+++ HookShield-frontend/Dockerfile
@@ -1,0 +1,23 @@
+# Use an official Node.js runtime as a parent image
+FROM node:20
+
+# Set the working directory
+WORKDIR /app
+
+# Copy package.json and package-lock.json
+COPY package*.json ./
+
+# Install dependencies
+RUN npm install
+
+# Copy the rest of the application code
+COPY . .
+
+# Build the application
+RUN npm run build
+
+# Expose the port the app runs on
+EXPOSE 4173
+
+# Command to run the application
+CMD ["npm", "run", "preview"]
