+++ HookShield-frontend/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview",
+    "preview": "vite preview --host",
     "test:unit": "vitest",
     "lint": "eslint . --fix"
   },
