+++ backend_api/.gitignore
@@ -1,6 +1,6 @@
 /venv
-/__pycache__
-/routers/__pycache__
-/db/__pycache__
-/utils/__pycache__
+app/__pycache__
+app/routers/__pycache__
+app/db/__pycache__
+app/utils/__pycache__
 pyvenv.cfg
