+++ backend_api/app/Dockerfile
@@ -1,0 +1,20 @@
+# Use an official Python image as a parent image
+FROM python:3.11-slim
+
+# Set the working directory in the container
+WORKDIR /app
+
+# Copy requirements file and install dependencies
+COPY requirements.txt .
+
+# Install dependencies
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copy the rest of the application code
+COPY . .
+
+# Expose the port your FastAPI app runs on
+EXPOSE 8000
+
+# Define the command to run your FastAPI app
+CMD ["uvicorn", "main:app", "--reload", "--host", "0.0.0.0", "--port", "8000"]
