--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "d3": "^7.9.0"
-  }
-}
